hand-rotated loops: InitGeometry's two log2 loops, GreyTable's u8 terms
InitGeometry 0x001619f0: the two `if (tileW > 1) { i32 v = tileW; do { v >>= 1;
m_shift++; } while (v > 1); }` blocks were cl's own rotation transcribed - retail
peels the guard itself out of a plain `while (v > 1)` and has NO preheader block.
Writing the natural form makes all 11 blocks agree instruction-for-instruction
(only the epilogue's register-save count still differs).  94.02 -> 95.89.

GreyTable 0x0014eef0: found while reading its loop bodies - retail casts ALL THREE
packed terms to u8 (`and dl,0xf` + `movzx dx,dl`), we cast only the first, so both
20-instruction loop bodies were 18.  Both bodies are now byte-exact.  81.05 -> 82.23;
the residue is one extra callee-saved register plus the known exit-merge family.

+++ src/DDrawMgr/ShadeTableCache.cpp
@@ -424,13 +424,15 @@ CShadeTable* CShadeTableCache::GreyTable() {
         && g_rUp == RGB555_RED_UP && g_gUp == PIXEL16_GREEN_UP) {
         for (i32 v = 0; v < 0x10000; v++) {
             *out++ = static_cast<u16>(
-                ((((static_cast<u8>((v >> 0xb)) << 4) + ((v >> 6) & 0xf)) << 4) + ((v >> 1) & 0xf))
+                ((((static_cast<u8>((v >> 0xb)) << 4) + static_cast<u8>((v >> 6) & 0xf)) << 4)
+                 + static_cast<u8>((v >> 1) & 0xf))
             );
         }
     } else {
         for (i32 v = 0; v < 0x10000; v++) {
             *out++ = static_cast<u16>(
-                ((((static_cast<u8>((v >> 0xc)) << 4) + ((v >> 7) & 0xf)) << 4) + ((v >> 1) & 0xf))
+                ((((static_cast<u8>((v >> 0xc)) << 4) + static_cast<u8>((v >> 7) & 0xf)) << 4)
+                 + static_cast<u8>((v >> 1) & 0xf))
             );
         }
     }
